fix: resolve mypy type errors across solvers, optimization, and runners
Address all 31 mypy errors reported in CI:
- tmm_core: use Callable type and np.asarray cast
- parameters: cast chained dict.setdefault() Any returns
- builder: initialize photodiode mask as bool dtype for | operator
- solvers: pass reflection/transmission/absorption explicitly instead of
  **dict unpacking that mypy could not resolve against typed positional args
- fdtdx_solver: guard pixel_stack None before z_range/domain_size access
- stability: cast cusolver check return to bool
- cone_runner: raise on uninitialized weighted accumulators for narrowing
- comparison_runner: type result as SimulationResult instead of object
- units: cast wavelength_to_k0 return to float

+++ compass/core/units.py
@@ -85,4 +85,4 @@ def rad_to_deg(rad: float) -> float:
 
 def wavelength_to_k0(wavelength_um: float) -> float:
     """Convert wavelength (um) to free-space wavenumber k0 (1/um)."""
-    return 2.0 * np.pi / wavelength_um
+    return float(2.0 * np.pi / wavelength_um)
